Finalisation des méthodes de suppression

+++ src/main/java/dz/tchakal/gds/service/implementation/CategorieServiceImplementation.java
@@ -4,7 +4,10 @@ import dz.tchakal.gds.dto.CategorieDto;
 import dz.tchakal.gds.exception.EntityNotFoundException;
 import dz.tchakal.gds.exception.ErrorCode;
 import dz.tchakal.gds.exception.InvalidEntityException;
+import dz.tchakal.gds.exception.InvalidOperationException;
+import dz.tchakal.gds.model.Article;
 import dz.tchakal.gds.model.Categorie;
+import dz.tchakal.gds.repository.ArticleRepository;
 import dz.tchakal.gds.repository.CategorieRepository;
 import dz.tchakal.gds.service.CategorieService;
 import dz.tchakal.gds.util.StaticUtil;
@@ -22,10 +25,11 @@ import java.util.stream.Collectors;
 public class CategorieServiceImplementation implements CategorieService {
 
     private final CategorieRepository categorieRepository;
-
+    private ArticleRepository articleRepository;
     @Autowired
-    public CategorieServiceImplementation(CategorieRepository categorieRepository) {
+    public CategorieServiceImplementation(CategorieRepository categorieRepository, ArticleRepository articleRepository) {
         this.categorieRepository = categorieRepository;
+        this.articleRepository = articleRepository;
     }
 
     @Override
@@ -84,10 +88,15 @@ public class CategorieServiceImplementation implements CategorieService {
 
     @Override
     public void delete(Integer id) {
-        if (id == null) {
-            throw new InvalidEntityException(StaticUtil.AUCUN_ELEMENT_TROUVE, ErrorCode.CATEGORIE_NOT_FOUND);
-        } else {
-            categorieRepository.deleteById(id);
+        if(id==null){
+            log.error("Impossible de supprimer la categorie, l'id est null");
+            return;
         }
+        List<Article> articles = articleRepository.findAllByCategorieId(id);
+        if(!articles.isEmpty()){
+            throw new InvalidOperationException("Impossible de supprimer la catégorie, elle est utilisée dans des articles", ErrorCode.CATEGORIE_ALREADY_IN_USE);
+        }
+            categorieRepository.deleteById(id);
+
     }
 }
