+++ src/main/java/dz/tchakal/gds/service/implementation/FournisseurServiceImplementation.java
@@ -4,7 +4,10 @@ import dz.tchakal.gds.dto.FournisseurDto;
 import dz.tchakal.gds.exception.EntityNotFoundException;
 import dz.tchakal.gds.exception.ErrorCode;
 import dz.tchakal.gds.exception.InvalidEntityException;
+import dz.tchakal.gds.exception.InvalidOperationException;
+import dz.tchakal.gds.model.CommandeFournisseur;
 import dz.tchakal.gds.model.Fournisseur;
+import dz.tchakal.gds.repository.CommandeFournisseurRepository;
 import dz.tchakal.gds.repository.FournisseurRepository;
 import dz.tchakal.gds.service.FournisseurService;
 import dz.tchakal.gds.util.StaticUtil;
@@ -22,10 +25,12 @@ import java.util.stream.Collectors;
 public class FournisseurServiceImplementation implements FournisseurService {
 
     private final FournisseurRepository fournisseurRepository;
+    private CommandeFournisseurRepository commandeFournisseurRepository;
 
     @Autowired
-    public FournisseurServiceImplementation(FournisseurRepository fournisseurRepository) {
+    public FournisseurServiceImplementation(FournisseurRepository fournisseurRepository, CommandeFournisseurRepository commandeFournisseurRepository) {
         this.fournisseurRepository = fournisseurRepository;
+        this.commandeFournisseurRepository = commandeFournisseurRepository;
     }
 
     @Override
@@ -72,9 +77,13 @@ public class FournisseurServiceImplementation implements FournisseurService {
     @Override
     public void delete(Integer id) {
         if (id == null) {
-            throw new InvalidEntityException(StaticUtil.AUCUN_ELEMENT_TROUVE, ErrorCode.ARTICLE_NOT_FOUND);
-        } else {
-            fournisseurRepository.deleteById(id);
+            log.error("Impossible de supprimer le client, l'id est null");
+            return;
         }
+        List<CommandeFournisseur> commandeFournisseurs = commandeFournisseurRepository.findAllByFournisseurId(id);
+        if (!commandeFournisseurs.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer le fournisseur, is est utilisé dans des commandes ", ErrorCode.FOURNISSEUR_ALREADY_IN_USE);
+        }
+        fournisseurRepository.deleteById(id);
     }
 }
